refact: 호스트 정보 저장 로직을 훅으로 분리하여 역할 분리
